INTRD-5212: merge into 12.X

+++ opencell-api/src/main/java/org/meveo/apiv2/catalog/service/PricePlanMatrixApiService.java
@@ -16,7 +16,6 @@ import java.util.Optional;
 
 import javax.inject.Inject;
 
-import org.apache.commons.codec.binary.Hex;
 import org.meveo.api.catalog.PricePlanMatrixLineApi;
 import org.meveo.api.dto.ActionStatusEnum;
 import org.meveo.api.dto.response.catalog.ImportResultResponseDto.ImportResultDto;
@@ -36,6 +35,7 @@ import org.meveo.jpa.MeveoJpa;
 import org.meveo.model.DatePeriod;
 import org.meveo.model.catalog.PricePlanMatrix;
 import org.meveo.model.catalog.PricePlanMatrixVersion;
+import org.meveo.model.cpq.enums.VersionStatusEnum;
 import org.meveo.model.shared.DateUtils;
 import org.meveo.service.catalog.impl.PricePlanMatrixColumnService;
 import org.meveo.service.catalog.impl.PricePlanMatrixService;
@@ -182,34 +182,39 @@ public class PricePlanMatrixApiService implements ApiService<PricePlanMatrix> {
                     pricePlanMatrixVersionService.update(ppmvToUpdate);
                 }
 
-                List<PricePlanMatrixVersion> previousPVs = pricePlanMatrixVersionService.findToDateAfterDateAndVersion(pricePlanMatrix, newFrom, ppmvToUpdate.getCurrentVersion());
-                for (PricePlanMatrixVersion previousPV : previousPVs) {
-                    if (previousPV.getId() != pricePlanVersionId && previousPV.getValidity().getFrom() != null && previousPV.getValidity().getFrom().compareTo(newFrom) > 0
-                            && previousPV.getValidity().getTo() != null && previousPV.getValidity().getTo().compareTo(newFrom) > 0) {
-                        pricePlanMatrixVersionService.remove(previousPV);
-                    } else {
-                        previousPV.setValidity(new DatePeriod(previousPV.getValidity().getFrom(), newFrom));
-                        pricePlanMatrixVersionService.update(previousPV);
-                        break;
-                    }
-                }
+                if (importItem.getStatus() == VersionStatusEnum.PUBLISHED) {
 
-                if (newTo != null) {
-                    List<PricePlanMatrixVersion> nextPVs = pricePlanMatrixVersionService.findFromDateBeforeDateAndVersion(pricePlanMatrix, newTo, ppmvToUpdate.getCurrentVersion());
-                    for (PricePlanMatrixVersion nextPV : nextPVs) {
-                        if (nextPV.getId() != pricePlanVersionId && nextPV.getValidity().getTo() != null && nextPV.getValidity().getTo().compareTo(newTo) < 0) {
-                            pricePlanMatrixVersionService.remove(nextPV);
+                    List<PricePlanMatrixVersion> previousPVs = pricePlanMatrixVersionService.findToDateAfterDateAndVersion(pricePlanMatrix, newFrom,
+                        ppmvToUpdate.getCurrentVersion());
+                    for (PricePlanMatrixVersion previousPV : previousPVs) {
+                        if (previousPV.getId() != pricePlanVersionId && previousPV.getValidity().getFrom() != null && previousPV.getValidity().getFrom().compareTo(newFrom) > 0
+                                && previousPV.getValidity().getTo() != null && previousPV.getValidity().getTo().compareTo(newFrom) > 0) {
+                            pricePlanMatrixVersionService.remove(previousPV);
                         } else {
-                            nextPV.setValidity(new DatePeriod(newTo, nextPV.getValidity().getTo()));
-                            pricePlanMatrixVersionService.update(nextPV);
+                            previousPV.setValidity(new DatePeriod(previousPV.getValidity().getFrom(), newFrom));
+                            pricePlanMatrixVersionService.update(previousPV);
                             break;
                         }
                     }
-                } else {
-                    List<PricePlanMatrixVersion> nextPVs = pricePlanMatrixVersionService.findFromDateAfterDateAndVersion(pricePlanMatrix, newFrom,
-                        ppmvToUpdate.getCurrentVersion());
-                    for (PricePlanMatrixVersion nextPV : nextPVs) {
-                        pricePlanMatrixVersionService.remove(nextPV);
+
+                    if (newTo != null) {
+                        List<PricePlanMatrixVersion> nextPVs = pricePlanMatrixVersionService.findFromDateBeforeDateAndVersion(pricePlanMatrix, newTo,
+                            ppmvToUpdate.getCurrentVersion());
+                        for (PricePlanMatrixVersion nextPV : nextPVs) {
+                            if (nextPV.getId() != pricePlanVersionId && nextPV.getValidity().getTo() != null && nextPV.getValidity().getTo().compareTo(newTo) < 0) {
+                                pricePlanMatrixVersionService.remove(nextPV);
+                            } else {
+                                nextPV.setValidity(new DatePeriod(newTo, nextPV.getValidity().getTo()));
+                                pricePlanMatrixVersionService.update(nextPV);
+                                break;
+                            }
+                        }
+                    } else {
+                        List<PricePlanMatrixVersion> nextPVs = pricePlanMatrixVersionService.findFromDateAfterDateAndVersion(pricePlanMatrix, newFrom,
+                            ppmvToUpdate.getCurrentVersion());
+                        for (PricePlanMatrixVersion nextPV : nextPVs) {
+                            pricePlanMatrixVersionService.remove(nextPV);
+                        }
                     }
                 }
 
@@ -268,10 +273,7 @@ public class PricePlanMatrixApiService implements ApiService<PricePlanMatrix> {
         if (StringUtils.isNotBlank(row)) {
             // Get the first character
             String bom = row.substring(0, 1);
-            // Convert first character to byte to character(Use Apache Commons Codec Hex class)
-            String bomByte = new String(Hex.encodeHex(bom.getBytes()));
-            if ("efbbbf".equals(bomByte)) {
-                // Eliminate BOM
+            if (!bom.equals("i")) { // i for id;
                 row = row.substring(1);
             }
         }
